move redoc theme to own file, use /docs route
